web api commands relocation to specific directories to improve dir layout
