+++ logstash-core/lib/logstash/api/lib/app/commands/stats/events_command.rb
