+++ logstash-core/lib/logstash/api/lib/app/commands/stats/hotthreads_command.rb
