+++ logstash-core/lib/logstash/api/lib/app/commands/stats/memory_command.rb
