+++ logstash-core/lib/logstash/api/lib/app/commands/system/basicinfo_command.rb
