+++ logstash-core/lib/logstash/api/lib/app/node_stats.rb
@@ -1,7 +1,5 @@
 # encoding: utf-8
 require "app"
-require "app/stats/events_command"
-require "app/stats/hotthreads_command"
 
 module LogStash::Api
   class NodeStats < BaseApp
