+++ logstash-core/lib/logstash/api/lib/app/root.rb
@@ -1,6 +1,5 @@
 # encoding: utf-8
 require "app"
-require "app/system/basicinfo_command"
 
 module LogStash::Api
   class Root < BaseApp
